+++ api/trips/trip-model.ts
@@ -1,17 +1,17 @@
 import { QueryBuilder } from 'knex';
 import db from '../../database/db-config';
-import { Trip, Person } from '../../types';
+import { Trip } from '../../types';
 
-interface TripUpdate {
+interface TripInput {
+  user_id: number;
   destination?: string;
   date?: Date;
   active?: boolean;
 }
-
-interface PersonInput {
-  first_name: string;
-  last_name?: string;
-  user_id?: string;
+interface TripUpdate {
+  destination?: string;
+  date?: Date;
+  active?: boolean;
 }
 
 const getByTripId = (id: number): QueryBuilder<{}, Trip> => {
@@ -28,20 +28,26 @@ const getByTripId = (id: number): QueryBuilder<{}, Trip> => {
     .first<Trip>();
 };
 
-const getPeopleByTripId = (id: number): QueryBuilder<{}, Person[]> => {
-  return db('people as p')
-    .where({ trip_id: id })
-    .join('trips as t', 't.id', 'p.trip_id')
-    .select(['p.id', 'p.first_name', 'p.last_name']);
+const getTripsByUsername = (username: string): Promise<Trip[]> => {
+  return db('trips as t')
+    .where({ username })
+    .join('users as u', 'u.id', 't.user_id')
+    .select(['t.id', 't.destination', 't.date', 't.active'])
+    .then(trips => {
+      return trips.map(trip => ({
+        ...trip,
+        active: Boolean(trip.active),
+      }));
+    });
 };
 
-const addPersonToTrip = (person: PersonInput, id: number): QueryBuilder => {
-  return db('people').insert(
+const addTripForUserId = (trip: TripInput, id: number): QueryBuilder => {
+  return db('trips').insert(
     {
-      trip_id: id,
-      first_name: person.first_name,
-      last_name: person.last_name,
-      user_id: person.user_id,
+      user_id: id,
+      destination: trip.destination || null,
+      date: trip.date || null,
+      active: trip.active || true,
     },
     'id'
   );
@@ -55,7 +61,7 @@ const updateTrip = (trip: TripUpdate, id: number): QueryBuilder<{}, Trip> => {
 
 export default {
   getByTripId,
-  getPeopleByTripId,
-  addPersonToTrip,
+  getTripsByUsername,
+  addTripForUserId,
   updateTrip,
 };
